@@ -2,13 +2,13 @@ namespace DesafioProjetoHospedagem.Models;
 
 public class Pessoa
 {
-    public Pessoa() { }
-
+    //Construtor com nome
     public Pessoa(string nome)
     {
         Nome = nome;
     }
 
+    //Construtor com nome e sobrenome
     public Pessoa(string nome, string sobrenome)
     {
         Nome = nome;
@@ -16,6 +16,7 @@ public class Pessoa
     }
 
     public string Nome { get; set; }
-    public string Sobrenome { get; set; }
+    public string? Sobrenome { get; set; }
+    //get: acesso somente leitura
     public string NomeCompleto => $"{Nome} {Sobrenome}".ToUpper();
 }
